% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid.R
\name{data_gen}
\alias{data_gen}
\title{Simulate dataset (used in Table 1 of Ye et al. (2021))}
\usage{
data_gen(n)
}
\arguments{
\item{n}{Sample size}
}
\value{
A data frame
}
\description{
Simulate dataset (used in Table 1 of Ye et al. (2021))
}
\examples{
df<-data_gen(1e5)
}
\references{
Ting Ye, Ashkan Ertefaie, James Flory, Sean Hennessy, and Dylan S. Small (2021). Instrumented Difference-in-Differences.
}
