% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid.R
\name{idid_g_comp}
\alias{idid_g_comp}
\title{G computation semiparametric estimator}
\usage{
idid_g_comp(yv, dv, tv, zv, X, V, fit_z, fit_t)
}
\arguments{
\item{yv}{A vector of outcome for each subject, missing is coded as NA}

\item{dv}{A vector of exposure for each subject, missing is coded as NA, length(dv)==length(yv)}

\item{tv}{A vector of time for each subject, missing is coded as NA, length(tv)==length(yv)}

\item{zv}{A vector of IV for DID for each subject, missing is coded as NA, length(zv)==length(yv)}

\item{X}{A matrix of observed covariates for each subject, missing is coded as NA, dim(X)[1]==length(yv)}

\item{V}{A matrix of effect modifiers for each subject, missing is coded as NA, dim(V)[1]==length(yv)}

\item{fit_z}{An glm object of Z using \code{X}}

\item{fit_t}{An glm object of T using \code{X}}
}
\value{
A list
\describe{
\item{est}{Estimated psi.}
\item{se}{Standard error of \code{est}.}
}
}
\description{
G computation semiparametric estimator
}
\details{
Here, delta(x) is linear in x and beta(v;psi) is linear in v. This function can be modified if one would like to use alternative nuisance parameter specifications.
}
\examples{
df<-data_gen(1e5)
fit_z<-glm(z~I(x1>0)+I(x2>0),data=df,family = "binomial") # one can use other models
fit_t<-glm(t~I(x1>0)+I(x2>0),data=df,family = "binomial") # one can use other models
X<-as.matrix(cbind(matrix(data=1,nrow = dim(df)[1],ncol=1),df[,grepl("x",names(df))]))
V<-matrix(data=1,nrow = dim(df)[1],ncol=1)
idid_g_comp(df$y,df$d,df$t,df$z,X,V,fit_z,fit_t)
V<-as.matrix(cbind(V,df[,grepl("x1",names(df))]))
idid_g_comp(df$y,df$d,df$t,df$z,X,V,fit_z,fit_t)
}
\references{
Ting Ye, Ashkan Ertefaie, James Flory, Sean Hennessy, and Dylan S. Small (2021). Instrumented Difference-in-Differences.
}
