% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid.R
\name{idid_reg}
\alias{idid_reg}
\title{Regression-based semiparametric estimator}
\usage{
idid_reg(yv, dv, tv, zv, X, V)
}
\arguments{
\item{yv}{A vector of outcome for each subject, missing is coded as NA}

\item{dv}{A vector of exposure for each subject, missing is coded as NA, length(dv)==length(yv)}

\item{tv}{A vector of time for each subject, missing is coded as NA, length(tv)==length(yv)}

\item{zv}{A vector of IV for DID for each subject, missing is coded as NA, length(zv)==length(yv)}

\item{X}{A matrix of observed covariates for each subject, missing is coded as NA, dim(X)[1]==length(yv)}

\item{V}{A matrix of effect modifiers for each subject, missing is coded as NA, dim(V)[1]==length(yv)}
}
\value{
A list
\describe{
\item{est}{Estimated psi.}
\item{se}{Standard error of \code{est}.}
}
}
\description{
Regression-based semiparametric estimator
}
\details{
Here, all the imposed nuisance parameters are linear in X. beta(v;psi) is also linear in v. This function can be modified if one would like to use alternative nuisance parameter specifications.
}
\examples{
df<-data_gen(1e5)
X<-as.matrix(cbind(matrix(data=1,nrow = dim(df)[1],ncol=1),df[,grepl("x",names(df))]))
V<-matrix(data=1,nrow = dim(df)[1],ncol=1)
idid_reg(df$y,df$d,df$t,df$z,X,V)
V<-as.matrix(cbind(V,df[,grepl("x1",names(df))]))
idid_reg(df$y,df$d,df$t,df$z,X,V)
}
\references{
Ting Ye, Ashkan Ertefaie, James Flory, Sean Hennessy, and Dylan S. Small (2021). Instrumented Difference-in-Differences.
}
