% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid.R
\name{idid_semi}
\alias{idid_semi}
\title{Instrumented DID semiparametric estimator}
\usage{
idid_semi(yv, dv, tv, zv, X, V, func_root, n.boot = 100)
}
\arguments{
\item{yv}{A vector of outcome for each subject, missing is coded as NA}

\item{dv}{A vector of exposure for each subject, missing is coded as NA, length(dv)==length(yv)}

\item{tv}{A vector of time for each subject, missing is coded as NA, length(tv)==length(yv)}

\item{zv}{A vector of IV for DID for each subject, missing is coded as NA, length(zv)==length(yv)}

\item{X}{A matrix of observed covariates for each subject, missing is coded as NA, dim(X)[1]==length(yv)}

\item{V}{A matrix of effect modifiers for each subject, missing is coded as NA, dim(V)[1]==length(yv)}

\item{func_root}{A function that calculates the root, e.g., idid_semi_root}

\item{n.boot}{Number of bootstrap interations}
}
\value{
\describe{
\item{psi.est}{Semiparametric estimators.}
\item{psi.se}{Standard error of \code{psi.est}.}
\item{f.statistic}{F statistic from the first-stage regression. It is preferably larger than 10.}
}
}
\description{
Instrumented DID semiparametric estimator
}
\examples{
df<-data_gen("case2",1e5)
V<-matrix(1,ncol=1,nrow=dim(df)[1])
idid_semi(df$y,df$d,df$t,df$z,matrix(df$x,ncol=1),V,idid_semi_root,n.boot = 10)
# n.boot is set to a small number for illustrative purpose
V<-cbind(matrix(1,ncol=1,nrow=dim(df)[1]),df$x)
idid_semi(df$y,df$d,df$t,df$z,matrix(df$x,ncol=1),V,idid_semi_root,n.boot = 10)
# n.boot is set to a small number for illustrative purpose
}
\references{
Ting Ye, Ashkan Ertefaie, James Flory, Sean Hennessy, and Dylan S. Small (2020). Instrumented Difference-in-Differences.
}
