% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid.R
\name{idid_semi_root}
\alias{idid_semi_root}
\title{Find the root based on the efficient influence function}
\usage{
idid_semi_root(yv, dv, tv, zv, X, V)
}
\arguments{
\item{yv}{A vector of outcome for each subject, missing is coded as NA}

\item{dv}{A vector of exposure for each subject, missing is coded as NA, length(dv)==length(yv)}

\item{tv}{A vector of time for each subject, missing is coded as NA, length(tv)==length(yv)}

\item{zv}{A vector of IV for DID for each subject, missing is coded as NA, length(zv)==length(yv)}

\item{X}{A matrix of observed covariates for each subject, missing is coded as NA, dim(X)[1]==length(yv)}

\item{V}{A matrix of effect modifiers for each subject, missing is coded as NA, dim(V)[1]==length(yv)}
}
\value{
A vector of roots
}
\description{
Find the root based on the efficient influence function
}
\details{
This function can be modified if one would like to use alternative working models.
}
\examples{
df<-data_gen("case2",1e5)
V<-matrix(1,ncol=1,nrow=dim(df)[1])
idid_semi_root(df$y,df$d,df$t,df$z,df$x,V)
V<-cbind(matrix(1,ncol=1,nrow=dim(df)[1]),df$x)
idid_semi_root(df$y,df$d,df$t,df$z,df$x,V)
}
\references{
Ting Ye, Ashkan Ertefaie, James Flory, Sean Hennessy, and Dylan S. Small (2020). Instrumented Difference-in-Differences.
}
