% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid.R
\name{idid_wald1}
\alias{idid_wald1}
\title{One-sample Instrumented DID Wald estimator}
\usage{
idid_wald1(yv, dv, tv, zv)
}
\arguments{
\item{yv}{A vector of outcome for each subject, missing is coded as NA}

\item{dv}{A vector of exposure for each subject, missing is coded as NA, length(dv)==length(yv)}

\item{tv}{A vector of time for each subject, missing is coded as NA, length(tv)==length(yv)}

\item{zv}{A vector of IV for DID for each subject, missing is coded as NA, length(zv)==length(yv)}
}
\value{
\describe{
\item{beta.est}{One-sample instrumented DID Wald estimate.}
\item{beta.se}{Standard error of \code{beta.est}.}
\item{f.statistic}{F statistic from the first-stage regression. It is preferably larger than 10.}
}
}
\description{
One-sample Instrumented DID Wald estimator
}
\examples{
df<-data_gen("case1",1e5)
idid_wald1(df$y,df$d,df$t,df$z)
}
\references{
Ting Ye, Ashkan Ertefaie, James Flory, Sean Hennessy, and Dylan S. Small (2020). Instrumented Difference-in-Differences.
}
