% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid.R
\name{data_gen}
\alias{data_gen}
\title{Simulate dataset}
\usage{
data_gen(case, n)
}
\arguments{
\item{case}{Simulation scenarios used in Tables 1-2 in Ye et al., (2020)}

\item{n}{Sample size}
}
\value{
A data frame
}
\description{
Simulate dataset
}
\examples{
df<-data_gen("case1",1e5)
}
\references{
Ting Ye, Ashkan Ertefaie, James Flory, Sean Hennessy, and Dylan S. Small (2020). Instrumented Difference-in-Differences.
}
